refactor: Move some things back to args.rs that aren't elf-specific
