+++ libwild/src/file_writer.rs
@@ -1,8 +1,8 @@
 use crate::OutputKind;
 use crate::args::Args;
+use crate::args::FileWriteMode;
 use crate::args::WRITE_VERIFY_ALLOCATIONS_ENV;
 use crate::args::elf::ElfArgs;
-use crate::args::elf::FileWriteMode;
 use crate::error;
 use crate::error::Context as _;
 use crate::error::Result;
