+++ libwild/src/grouping.rs
@@ -203,11 +203,11 @@ fn determine_symbols_per_group(num_symbols: usize, args: &Args<ElfArgs>) -> usiz
     // If we have lots of threads, then we might benefit from a few more groups in order to properly
     // take advantage of the available parallelism.
     let groups_per_thread =
-        args.numeric_experiment(crate::args::elf::Experiment::GroupsPerThread, 5) as usize;
+        args.numeric_experiment(crate::args::Experiment::GroupsPerThread, 5) as usize;
 
     // If we don't have lots of threads, then we still want a reasonable number of groups. The need
     // for this was based on experimentation.
-    let min_groups = args.numeric_experiment(crate::args::elf::Experiment::MinGroups, 150) as usize;
+    let min_groups = args.numeric_experiment(crate::args::Experiment::MinGroups, 150) as usize;
 
     let target_num_groups = (num_threads * groups_per_thread).max(min_groups);
